structure change to MyLibrary, using AddBook component
