+++ client/components/BorrowedBooks.js
@@ -1,3 +1,14 @@
+/**
+ * ************************************
+ *
+ * @module  BorrowedBooks.js
+ * @author
+ * @date
+ * @description display componenent for user library books borrowed from others
+ *
+ * ************************************
+ */
+
 import React from 'react'
 
 const BorrowedBooks = () => {
