+++ client/components/CheckedOut.js
@@ -1,3 +1,14 @@
+/**
+ * ************************************
+ *
+ * @module  CheckedOut.js
+ * @author
+ * @date
+ * @description display componenent for user library books that are checked out
+ *
+ * ************************************
+ */
+
 import React from 'react'
 
 const CheckedOut = () => {
